two maps, one for localization one for navigation
We've recorded a new map for the project room with the big laser scanner (hokuyo_30lx).
In this case we decided to split the maps in two parts. Move base uses the navigation map,
all other parts uses the localization map, therefore this map keeps the name topic name map.