Fix gmapping laser scan offset and max range error
The minimum and maximum angle was not set correctly in the driver for the
Hokuyo 04lx. This lead to an offset in gmapping. Now, use different
values for the 04lx and the 30lx.

Also, the maxRange parameter was set to a higher value, than the max range
of the Hokuyo 04lx actually is. So, the max range values of the 04lx were
used to build the map.
We do not set that parameter anymore, so that gmapping uses the values
from the laser scan message.